Move vote logic from post controller to vote controller
