+++ server/controllers/vote.controller.js
@@ -1,3 +1,7 @@
+import raccoon from 'raccoon';
+import Bluebird from 'bluebird';
+import mongoose from 'mongoose';
+
 import Vote from '../models/vote.model';
 
 /**
@@ -77,4 +81,123 @@ function remove(req, res, next) {
     .catch(e => next(e));
 }
 
-export default { load, get, create, update, list, remove };
+/**
+ * Upvote a post.
+ */
+function upvote(req, res, next) {
+  const post = req.post;
+
+  if (!post.score) post.score = 0;
+
+  Vote.findOne({
+    postId: post._id,
+    userId: req.user._id,
+  })
+  .then((voteFound) => {
+    const vote = voteFound;
+    const userIdString = req.user._id.toString();
+    const postIdString = post._id.toString();
+
+    if (vote) {
+      let incrementValue = 1;
+
+      // We are changing directly from down to up
+      if (vote.direction !== 'upvote' && vote.active) {
+        incrementValue = 2;
+      }
+
+      vote.active = !vote.active;
+
+      if (vote.direction !== 'upvote') {
+        vote.direction = 'upvote';
+        vote.active = true;
+      }
+
+      if (vote.active) {
+        post.score += incrementValue;
+        raccoon.liked(userIdString, postIdString);
+      } else {
+        post.score -= incrementValue;
+        raccoon.unliked(userIdString, postIdString);
+      }
+
+      return Bluebird.all([vote.save(), post.save()]);
+    }
+
+    const newvote = new Vote();
+    newvote.postId = post._id;
+    newvote.userId = req.user._id;
+    newvote.direction = 'upvote'; // @TODO: Make constant
+    post.score += 1;
+    raccoon.liked(userIdString, postIdString);
+
+    return Bluebird.all([newvote.save(), post.save()]);
+  })
+  .then((vote) => {
+    req.vote = vote[0]; // eslint-disable-line no-param-reassign
+    return res.json(vote[0]);
+  })
+  .catch((e) => {
+    next(e);
+  });
+}
+
+
+/**
+ * Downvote a post.
+ */
+function downvote(req, res, next) {
+  const post = req.post;
+
+  if (!post.score) post.score = 0;
+
+  Vote.findOne({
+    postId: post._id,
+    userId: req.user._id,
+  })
+  .then((voteFound) => {
+    const vote = voteFound;
+    if (vote) {
+      let incrementValue = 1;
+
+      // We are changing directly from up to down
+      if (vote.direction !== 'downvote' && vote.active) {
+        incrementValue = 2;
+      }
+
+      vote.active = !vote.active;
+
+      if (vote.direction !== 'downvote') {
+        vote.direction = 'downvote';
+        vote.active = true;
+      }
+
+      if (vote.active) {
+        post.score -= incrementValue;
+        raccoon.disliked(req.user._id.toString(), post._id.toString());
+      } else {
+        post.score += incrementValue;
+        raccoon.undisliked(req.user._id.toString(), post._id.toString());
+      }
+
+      return Bluebird.all([vote.save(), post.save()]);
+    }
+
+    const newvote = new Vote();
+    newvote.postId = post._id;
+    newvote.userId = req.user._id;
+    newvote.direction = 'downvote'; // @TODO: Make constant
+    post.score -= 1;
+
+    raccoon.disliked(req.user._id.toString(), post._id.toString());
+
+    return Bluebird.all([newvote.save(), post.save()]);
+  })
+  .then((vote) => {
+    req.vote = vote[0]; // eslint-disable-line no-param-reassign
+    return res.json(vote[0]);
+  })
+  .catch(e => next(e));
+}
+
+export default { load, get, create, update, list, remove, upvote, downvote };
